moved login functionality to a component
